[FEATURE] add Flower Specification

+++ src/main/java/com/enigma/sun_florist/dto/request/SearchFlowerRequest.java
@@ -22,7 +22,7 @@ public class SearchFlowerRequest {
 
     private Integer minStock;
 
-    private Integer maxPrice;
+    private Long maxPrice;
 
-    private Integer minPrice;
+    private Long minPrice;
 }
